Fixed warnings from chat and library Cocoa demos

+++ objective-c/IceTouch/Cocoa/chat/Classes/ChatController.m
@@ -259,7 +259,7 @@
 
 #pragma mark NSTableView delegate
 
-- (int)numberOfRowsInTableView:(NSTableView *)tv 
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tv
 { 
     return users.count;
 }
